delete method implementation.
