+++ hw10-modern-web/src/main/resources/application.yml
@@ -1,7 +1,5 @@
 spring:
   data:
-    rest:
-      base-path: /api
     mongodb:
       host: localhost
       port: 27017
